Changed highestCost data type to int in GetSortedProducts_Valid_Cost_Selected_Should_Return_Cities_Sorted_By_Cost unit test in Productlist.Razor.Tests.cs

+++ UnitTests/Components/ProductList.razor.Tests.cs
@@ -738,7 +738,7 @@ namespace UnitTests.Components
             var productModels = TestHelper.ProductService.GetAllData();
 
             // The highest cost in the database
-            double highestCost = productModels.Max(x => x.Cost);
+            int highestCost = productModels.Max(x => x.Cost);
 
             // The lowest cost in the database
             double lowestCost = productModels.Min(x => x.Cost);
